added texture pattern
